disable explicit MPI finalization
MPI.jl will handle this at shutdown and by waiting,
we allow other parts of the julia session to use MPI.
In the future, MPI will bring features for scoped
(de)initialization, but for now that is it.

A drawback is that in case of a bug, MPI can be left
in an unclean state and the next MPI code run in the
same session will deadlock.