scheduler_mpi: streamlining protocol
Get rid of the concurrent work-sharing controller. The state
of concurrency in Julia makes it very difficult to have the
controller process unoccupied without putting restrictions on the user.

The new protocol merges several messages into one and avoids
the controller having to wait for workers comms to synchronize
among themselves.